Remove api-request
