+++ api/index.js
@@ -1,8 +1,6 @@
 'use strict';
 
 var SuiteRequest = require('escher-suiteapi-js');
-var SuiteRequestOptions = SuiteRequest.Options;
-var ApiRequest = require('./../lib/api-request');
 var InternalApiRequest = require('./../lib/internal-api-request');
 var AdministratorAPI = require('./endpoints/administrator');
 var ContactAPI = require('./endpoints/contact');
@@ -53,14 +51,6 @@ SuiteAPI.prototype = {
   _createInternalApiRequest: function(options) {
     return InternalApiRequest.create(options);
   },
-
-
-  _createServiceApiRequest: function(options) {
-    var requestOptions = SuiteRequestOptions.createForServiceApi(options.environment, options.rejectUnauthorized);
-    var suiteRequest = SuiteRequest.create(options.apiKey, options.apiSecret, requestOptions);
-    return ApiRequest.create(options);
-  },
-
 
   _apiKeySecret: function() {
     var apiKey = process.env.SUITE_API_KEY;
