+++ lib/internal-api-request/index.spec.js
@@ -207,7 +207,7 @@ describe('InternalApiRequest', function() {
 
   var getPromiseResolvesWith = function(respObj) {
     return new Promise(function(resolve) {
-      resolve({ data: { data: respObj } });
+      resolve({ data: respObj });
     });
   };
